SynchronizedWorkerThread: Fix possible race-condition between thread.
The blocking waitForState call has to give up the lock for the thread to
take, so that they don't touch the state at the same time.
However, this allows another call to waitForState or moveToState to take
the lock as well, moving the thread into a different state and causing
the original waitForState invocation to block until its requested state
comes along again.

This has never been a problem because Android should not request
multiple state transitions at the same time, but better be safe than
sorry.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>